advisor-backend/src/topic/topic.module.ts
import { Module } from '@nestjs/common';
import { TopicService } from './topic.service';
import { TopicController } from './topic.controller';

// NestJS Module System helps structure application code into domain areas
//  leading to better design and improved maintainability.
// Controllers are responsible for handling incoming requests and returning responses to the client.
// Providers are responsible for executing business logic and returning data to the controller.
// Module for topic related routes
@Module({
  controllers: [TopicController],
  providers: [TopicService],
})
export class TopicModule {}
